requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==1.6.3
Babel==2.5.3
billiard==3.3.0.23
celery==3.1.18
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
Django==1.10.2
django-appconf==1.0.2
django-bootstrap-form==3.4
django-celery==3.2.2
django-paypal==0.5.0
django-user-accounts==2.0.3
flower==0.9.2
gunicorn==19.7.1
idna==2.6
isort==4.3.4
kombu==3.0.37
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pylint==1.8.4
pytz==2018.3
redis==2.10.6
requests==2.18.4
six==1.11.0
South==1.0.2
tornado==5.0.1
urllib3==1.22
whitenoise==3.3.1
wrapt==1.10.11
